requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
civis==1.11.0
Click==7.0
cloudpickle==1.2.2
facebook-sdk==3.1.0
idna==2.8
joblib==0.13.2
jsonref==0.2
jsonschema==3.0.0
pubnub==4.1.6
pycryptodomex==3.9.3
pyrsistent==0.15.5
PyYAML==5.1.2
requests==2.22.0
retrying==1.3.3
six==1.13.0
urllib3==1.25.7
